consolidate helpers into one
