+++ .github/workflows/update-merge-queue.yml
@@ -3,11 +3,11 @@ name: Update Merge Queue
 on:
   pull_request:
     branches: [ main ]
-    paths:
-      - 'src/helpers/update-merge-queue.ts'
+    types: [ closed ]
 
 jobs:
   test:
+    if: github.event.pull_request.merged == true
     runs-on: ubuntu-latest
     steps:
       - name: Checkout
@@ -15,5 +15,5 @@ jobs:
 
       - uses: ./
         with:
-          helper: update-merge-queue
+          helper: manage-merge-queue
           github_token: ${{ secrets.GITHUB_TOKEN }}
