+++ test/helpers/manage-merge-queue.test.ts
@@ -12,30 +12,42 @@ limitations under the License.
 */
 
 import { Mocktokit } from '../types';
-import { addPrToMergeQueue } from '../../src/helpers/add-pr-to-merge-queue';
+import { addLabels } from '../../src/helpers/add-labels';
 import { context } from '@actions/github';
+import { manageMergeQueue } from '../../src/helpers/manage-merge-queue';
 import { octokit } from '../../src/octokit';
+import { removeLabel } from '../../src/helpers/remove-label';
 
 jest.mock('@actions/core');
 jest.mock('@actions/github', () => ({
   context: { repo: { repo: 'repo', owner: 'owner' }, issue: { number: 123 } },
   getOctokit: jest.fn(() => ({
     rest: {
-      issues: { addLabels: jest.fn() },
+      issues: { addLabels: jest.fn(), listLabelsOnIssue: jest.fn() },
       search: { issuesAndPullRequests: jest.fn() }
     }
   }))
 }));
 (octokit.search.issuesAndPullRequests as unknown as Mocktokit).mockImplementation(async () => ({
   data: {
-    total_count: 3
+    total_count: 3,
+    items: [
+      {
+        number: 123,
+        labels: [{ name: 'QUEUED FOR MERGE #2' }]
+      },
+      {
+        number: 456,
+        labels: [{ name: 'QUEUED FOR MERGE #3' }]
+      }
+    ]
   }
 }));
 
 describe('addPrToMergeQueue', () => {
   const sha = 'sha';
   beforeEach(async () => {
-    await addPrToMergeQueue({ sha });
+    await manageMergeQueue({ sha });
   });
 
   it('should call issuesAndPullRequests search with correct params', () => {
@@ -49,6 +61,28 @@ describe('addPrToMergeQueue', () => {
       labels: ['QUEUED FOR MERGE #4'],
       issue_number: 123,
       ...context.repo
+    });
+  });
+
+  it('should call add labels with correct params', () => {
+    expect(addLabels).toHaveBeenCalledWith({
+      pull_number: '123',
+      labels: 'QUEUED FOR MERGE #1'
+    });
+    expect(addLabels).toHaveBeenCalledWith({
+      pull_number: '456',
+      labels: 'QUEUED FOR MERGE #2'
+    });
+  });
+
+  it('should call remove label with correct params', () => {
+    expect(removeLabel).toHaveBeenCalledWith({
+      pull_number: '123',
+      label: 'QUEUED FOR MERGE #2'
+    });
+    expect(removeLabel).toHaveBeenCalledWith({
+      pull_number: '456',
+      label: 'QUEUED FOR MERGE #3'
     });
   });
 });
